langchain-demos/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
pydantic>=2.0.0
python-dotenv>=1.0.0